requirements.txt
python-telegram-bot==22.6
python-dotenv==1.2.1
pytest>=9.0
